+++ src/main/java/com/sprint/mission/discodeit/exception/binaryContent/BinaryContentNotFoundException.java
@@ -1,11 +1,17 @@
-package com.sprint.mission.discodeit.exception.binaryContent;
+package com.sprint.mission.discodeit.exception.binarycontent;
 
-import com.sprint.mission.discodeit.exception.DiscodeitException;
 import com.sprint.mission.discodeit.exception.ErrorCode;
 
-public class BinaryContentNotFoundException extends DiscodeitException {
+import java.util.UUID;
 
-    public BinaryContentNotFoundException(ErrorCode errorCode) {
-        super(errorCode);
+public class BinaryContentNotFoundException extends BinaryContentException {
+    public BinaryContentNotFoundException() {
+        super(ErrorCode.BINARY_CONTENT_NOT_FOUND);
+    }
+
+    public static BinaryContentNotFoundException withId(UUID binaryContentId) {
+        BinaryContentNotFoundException exception = new BinaryContentNotFoundException();
+        exception.addDetail("binaryContentId", binaryContentId);
+        return exception;
     }
 }
